💄 Style: sideBar 스타일 수정
