+++ src/assets/icons/index.ts
@@ -18,3 +18,4 @@ export { default as IconSetting } from './setting.svg?react';
 export { default as IconCopyLink } from './copyLink.svg?react';
 export { default as IconCamera } from './camera.svg?react';
 export { default as IconImageGallery } from './imageGallery.svg?react';
+export { default as IconTopArrow } from './topArrow.svg?react';
