+++ src/pages/Diary/components/DiaryCommon/index.ts
@@ -1,2 +1,3 @@
 export { default as DiaryCategoryList } from './DiaryCategoryList';
 export { default as DiaryHeader } from './DiaryHeader';
+export { default as DiarySideBar } from './DiarySideBar';
